requirements.txt
uvicorn
pydantic
pymongo
bcrypt
python-dotenv
email-validator
passlib
sendgrid
secure-smtplib
motor
pyjwt
pyarabic
python-bidi
arabic-reshaper
